refactor(env): Renames scripts to management to make it consistent
