+++ package.json
@@ -5,13 +5,13 @@
   "main": "index.js",
   "scripts": {
     "lint": "eslint src test",
-    "test": "rimraf .nyc_output && NODE_ENV=test WT_CONFIG=test ./test/utils/scripts/localtestnet.sh",
+    "test": "rimraf .nyc_output && NODE_ENV=test WT_CONFIG=test ./test/utils/management/localtestnet.sh",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "start": "node ./src/index.js",
     "dev": "WT_CONFIG=dev npm start",
-    "deploy-aws-playground": "./scripts/deploy-aws.sh playground",
-    "deploy-aws-demo": "./scripts/deploy-aws.sh demo",
-    "dev-net": "./scripts/dev-net.sh"
+    "deploy-aws-playground": "./management/deploy-aws.sh playground",
+    "deploy-aws-demo": "./management/deploy-aws.sh demo",
+    "dev-net": "./management/dev-net.sh"
   },
   "repository": {
     "type": "git",
